// packages/diagrams/frontend/sirius-components-diagrams-reactflow/src/renderer/fit-to-screen/useInitialFitToScreen.types.ts
/*******************************************************************************
 * Copyright (c) 2023 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/

export interface UseInitialFitToScreenValue {
  fitToScreen: () => void;
}

export interface UseInitialFitToScreenState {
  initialFitToScreenPerformed: boolean;
  shouldPerformFitToScreen: boolean;
}